+++ @planetarium/tx/src/tx/index.ts
@@ -1,4 +1,10 @@
 export type { TxMetadata } from "./metadata";
 export { encodeTxMetadata } from "./metadata";
-export type { UnsignedTxWithSystemAction } from "./unsigned";
-export { encodeUnsignedTxWithSystemAction } from "./unsigned";
+export type {
+  UnsignedTxWithSystemAction,
+  UnsignedTxWithCustomActions,
+} from "./unsigned";
+export {
+  encodeUnsignedTxWithSystemAction,
+  encodeUnsignedTxWithCustomActions,
+} from "./unsigned";